requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.1.0
python-dotenv==1.0.1
pyodbc
pillow==10.2.0 